refactor(api): 💡 extract common sqlite functions to util module
